requirements.txt
# weaviate-client @ git+https://github.com/weaviate/weaviate-python-client.git@dev/1.26
weaviate-client
numpy
semver
click
